Saco Id de mas en User

+++ src/main/java/controllers/ItemsController.java
@@ -37,7 +37,7 @@ public class ItemsController extends WebApiController{
 //        Listing listing = this.listingsApi.getListing("MLA527664161");
 
         Item item = new Item(user, listing.description, listing.picture);
-        int id = this.itemHome.create(item);
+        long id = this.itemHome.create(item);
 
         return Results.json().render(id);
     }
